+++ test/AdaptiveMPP.jl
@@ -21,7 +21,7 @@ data    = [ 'a', 'b', 'c', 'c', 'd',
 end
 
 # Test model
-test_model = Dict{Vector{Char},Int64}(
+test_model = Dict{Vector{Char},Int}(
     ['a'] => 1,
     ['a','b'] => 1,
     ['a','b','c'] => 1,
@@ -80,7 +80,7 @@ end
     # Reusing p
     # @test get_best_symbol( p ) == 'b'
     # For untrained model
-    @test get_best_symbol( AdaptiveMPP{Int64}(4) ) == nothing
+    @test get_best_symbol( AdaptiveMPP{Int}(4) ) == nothing
     @test get_best_symbol( AdaptiveMPP{Char}(4) ) == nothing
     @test get_best_symbol( AdaptiveMPP{String}(4) ) == nothing
 end
@@ -92,9 +92,9 @@ end
 
 @testset "Random Sequence test" begin
     for idx = 1:10
-        learnt_model    = AdaptiveMPP{Int64}(5);
+        learnt_model    = AdaptiveMPP{Int}(5);
         @test @test_nothrow for i = 1:1000
-            symbol = trunc(Int64,10*rand());
+            symbol = trunc(Int,10*rand());
             add!( learnt_model, symbol );
             predict( learnt_model );
         end
