+++ test/LeZiUpdate.jl
@@ -24,7 +24,7 @@ data    = [ 'a', 'a', 'a', 'b', 'a',
 end
 
 # Test model
-test_model = Dict{Vector{Char},Int64}(
+test_model = Dict{Vector{Char},Int}(
     ['a'] => 10,
     ['a','a'] => 3,
     ['a','a','a'] => 1,
@@ -86,9 +86,9 @@ end
 
 @testset "Random Sequence test" begin
     for idx = 1:10
-        learnt_model    = LeZiUpdate{Int64}();
+        learnt_model    = LeZiUpdate{Int}();
         @test @test_nothrow for i = 1:1000
-            symbol = trunc(Int64,10*rand());
+            symbol = trunc(Int,10*rand());
             add!( learnt_model, symbol );
             predict( learnt_model );
         end
